Pure Java compile & test example with Leiningen.
It took entirely too much fiddling and work to make this happen. As
I'm a lazy bum who finds Maven, Ant, and other JVM-based build systems
to be equivalent to setting your hair on fire... lein is ideal! And
Lispy, hurray.

Anyway, its example code, licenced as CC0 for anyone to muck with,
fork, etc. Enjoy!